fix(name): déplacement du js en global
