+++ src/component/main.js
@@ -12,9 +12,7 @@ import './navigation/main.js';
 import './tab/main.js';
 import './table/main.js';
 import './tag/main.js';
-import './download/main.js';
 import './transcription/main.js';
-import './tooltip/main.js';
 import './tile/main.js';
 import './header/main.js';
 import './display/main.js';
