--- src/pattern/name/example/data/js/remove-firstname.js
@@ -1,3 +1,0 @@
-const input = document.getElementById('<%= id %>-added-${this.firstnamesCount}');
-input.closest('.<%= prefix %>-fieldset__element').remove();
-this.closest('.<%= prefix %>-fieldset__element').remove();
