--- src/pattern/name/example/data/js/toggle-disable.js
@@ -1,3 +1,0 @@
-const fieldset = document.getElementById('<%= id %>');
-if (this.checked) fieldset.setAttribute('disabled', '');
-else fieldset.removeAttribute('disabled');
