+++ src/pattern/name/example/js/toggle-disabled.js
@@ -1,0 +1,5 @@
+window.toggleDisabled = (checkbox, id) => {
+  const fieldset = document.getElementById(id);
+  if (checkbox.checked) fieldset.setAttribute('disabled', '');
+  else fieldset.removeAttribute('disabled');
+};
